record-to-pycsw/requirements.txt
rdflib
psycopg2-binary
SQLAlchemy<2.0.0
python-dotenv
# git+https://github.com/geopython/pycsw.git@master#egg=pycsw # use this locally, not on git CI
